function/requirements.txt
azure-functions==1.23.0
azure-cosmos==4.9.0
validators==0.35.0
